Add files via upload
1. token.py
This code (class Token) stores the tokens that come from the Talismane parsed files.

2. contexts.py
This code (class Context) reads each Talismane parsed file, identifies the tokens, counts how many sentences there are in the context and verifies the presence of a citation.

3. features_module.py
This module presents a collection of funcitons to identify different negative features present in the contexts. The presence of the features is indicated by a boolean.

4. rules_module.py
This module presents a collection of funcitons to identify the patterns that constitute the different rules. The presence of the rule patters is indicated by a boolean.

5. evaluator.py
This function (class Evaluator) stores the results from the features and rules modules in a dictionary which takes the id of the context as key. ALso, the text of the context and the original polarity, both coming from the CitaNeg dataset file, are stored in this dictionary. The text of the context is useful to analyze the rules result and the original polarity to verify precision and recall.
This class also applies the rules and features and evaluates their results (result of a single rule/feature or of a collection of rules/features). 
Finaly, the class exports the results under csv format.

6. main.py
The main code opens the directory where the Talismane parsed files are stored. It allow to apply the features/rules and evaluate them by calling the class Evaluator.